+++ app/javascript/init/unwrap_moment.js
@@ -1,1 +1,0 @@
-window.moment = modulejs.require('moment');