+++ hbw/app/javascript/packs/hbw/components/shared/element/date_picker_wrapper.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import DatePicker from 'react-datepicker';
+import 'react-datepicker/dist/react-datepicker.css';
+
+const DatePickerWrapper = props => <DatePicker {...props}/>;
+
+export default DatePickerWrapper;