+++ hbw/app/javascript/packs/hbw/components/shared/utils/to_fns_date_format.js
@@ -1,0 +1,5 @@
+// date-fns uses 'd' and 'y' for month days and calendar years
+// as opposed to moment.js that uses capital 'D' and 'Y'
+const convertToDateFNSDateFormat = dateFormat => dateFormat.replaceAll('Y', 'y').replaceAll('D', 'd');
+
+export default convertToDateFNSDateFormat;