+++ spec/hbw/features/bp_form/description_date_time_spec.rb
@@ -15,11 +15,11 @@ feature 'Check description on top and botton of datetime', js: true do
 
   scenario 'should contain a top and bottom datetime descriptions' do
     parent_top = page.find('.datetime-top')
-    expect(page).to have_selector "[name='topDescriptionDateTime']"
+    expect(page).to have_selector "[name='topDescriptionDateTime-visible-input']"
     expect(find_by_dt('description-top', parent_top).text).to eq 'Top description test'
 
     parent_bottom = page.find('.datetime-bottom')
-    expect(page).to have_selector :css, "input[name='bottomDescriptionDatetime']"
+    expect(page).to have_selector :css, "input[name='bottomDescriptionDatetime-visible-input']"
     expect(find_by_dt('description-bottom', parent_bottom).text).to eq 'Bottom description test'
   end
 end