+++ spec/hbw/features/bp_form/dynamic_conditions_by_select_table_spec.rb
@@ -64,7 +64,7 @@ feature 'Control fields with dynamic conditions by select table', js: true do
     expect(readonly?('dependentString')).to        be false
     expect(readonly?('dependentText')).to          be false
     expect(readonly?('stringInDependentGroup')).to be false
-    expect(find_by_name('dependentDatetime').disabled?).to    be false
+    expect(find_by_name('dependentDatetime-visible-input').disabled?).to be false
     expect(radio_button_disabled?('dependentRadioButton')).to be false
 
     in_group_with_label('Dependent submit_select group') do |group|
@@ -142,7 +142,7 @@ feature 'Control fields with dynamic conditions by select table', js: true do
     expect(readonly?('dependentString')).to        be false
     expect(readonly?('dependentText')).to          be false
     expect(readonly?('stringInDependentGroup')).to be false
-    expect(find_by_name('dependentDatetime').disabled?).to    be false
+    expect(find_by_name('dependentDatetime-visible-input').disabled?).to be false
     expect(radio_button_disabled?('dependentRadioButton')).to be false
 
     in_group_with_label('Dependent submit_select group') do |group|
@@ -164,7 +164,7 @@ feature 'Control fields with dynamic conditions by select table', js: true do
     expect(readonly?('dependentString')).to        be true
     expect(readonly?('dependentText')).to          be true
     expect(readonly?('stringInDependentGroup')).to be true
-    expect(find_by_name('dependentDatetime').disabled?).to    be true
+    expect(find_by_name('dependentDatetime-visible-input').disabled?).to be true
     expect(radio_button_disabled?('dependentRadioButton')).to be true
 
     in_group_with_label('Dependent submit_select group') do |group|