+++ CHANGELOG.md
@@ -15,6 +15,9 @@ v2.7.0 [unreleased]
 - [#586](https://github.com/latera/homs/pull/586) Stub api container in tests with local api mock file and refactor tests.
 - [#594](https://github.com/latera/homs/pull/594) Get rid of Webpacker & build all assets with Webpack 5.
 - [#598](https://github.com/latera/homs/pull/598) Remove the mask option validation from the string order field.
+- [#600](https://github.com/latera/homs/pull/600) Replace bootstrap-datetimepicker with react-datepicker.
+
+  Dynamic conditions by datetime on form now strictly use ISO date format
 
 v2.6.14 [unreleased]
 -------------------